UI changed make web3auth will use for login only and after that user can use solana and EVM both in single app
