+++ package.json
@@ -22,7 +22,7 @@
     "@web3auth/modal-react-hooks": "^9.5.1",
     "@web3auth/openlogin-adapter": "^8.12.4",
     "@web3auth/web3auth": "^2.1.3",
-    "next": "15.1.7",
+    "next": "^15.3.0-canary.6",
     "pino-pretty": "^13.0.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
